+++ include/mondata.h
@@ -232,6 +232,10 @@
         || ((ptr) == &mons[PM_LEMURE] \
             || (ptr) == &mons[PM_MANES] \
             || (ptr) == &mons[PM_HELLCAT]))
+#define is_ghoul(ptr) ((ptr) == &mons[PM_GHOUL] \
+                       || (ptr) == &mons[PM_GHAST] \
+                       || (ptr) == &mons[PM_GHOUL_MAGE] \
+                       || (ptr) == &mons[PM_VOMITOUS_GHOUL])
 #define is_minion(ptr) (((ptr)->mflags2 & M2_MINION) != 0L)
 #define likes_gold(ptr) (((ptr)->mflags2 & M2_GREEDY) != 0L)
 #define likes_gems(ptr) (((ptr)->mflags2 & M2_JEWELS) != 0L)
@@ -376,6 +380,7 @@
 
 #define avoids_player(ptr) \
     (is_unicorn(ptr) \
+     || (ptr) == &mons[PM_ALCHEMIST] \
      || (ptr) == &mons[PM_MAGICAL_EYE])
 
 #define horizontal_mover(ptr) \
@@ -386,7 +391,8 @@
 
 #define is_bones_monster(ptr) ((ptr) == &mons[PM_GHOST] || (ptr) == &mons[PM_GHOUL]        \
                                || (ptr) == &mons[PM_VAMPIRE] || (ptr) == &mons[PM_WRAITH]  \
-                               || (ptr) == &mons[PM_GREEN_SLIME] || (ptr)->mlet == S_MUMMY)
+                               || (ptr) == &mons[PM_GREEN_SLIME] || (ptr)->mlet == S_MUMMY \
+                               || (ptr) == &mons[PM_BAOBHAN_SITH])
 /* Noise that a monster makes when engaged in combat. Assume that vocalizations
  * account for some noise, so monsters capable of vocalizing make more.
  * This gets used as an argument to wake_nearto, which expects a squared value,