+++ src/mhitm.c
@@ -377,6 +377,9 @@ mattackm(register struct monst *magr, register struct monst *mdef)
     if (magr->data == &mons[PM_HYDRA]) {
         k = min(magr->m_lev - magr->data->mlevel + 1, 10);
     }
+    if (magr->data == &mons[PM_HECATONCHEIRE]) {
+        k = 100;
+    }
 
     /* controls whether a mind flayer uses all of its tentacle-for-DRIN
        attacks; when fighting a headless monster, stop after the first
@@ -588,6 +591,11 @@ mattackm(register struct monst *magr, register struct monst *mdef)
 
         /* handle multiple hydra attacks */
         if (magr->data == &mons[PM_HYDRA] && mattk->aatyp == AT_BITE && k > 0) {
+            i -= 1;
+            k -= 1;
+        }
+        /* hundred handed ones attack until they miss (up to 100 times) */
+        if (magr->data == &mons[PM_HECATONCHEIRE] && (res[i] & MM_HIT)) {
             i -= 1;
             k -= 1;
         }