+++ src/pager.c
@@ -902,6 +902,11 @@ struct permonst * pm;
         else if (vegetarian(pm))
             MONPUTSTR("May be eaten by vegetarians.");
     }
+    if (pm == &mons[PM_HECATONCHEIRE]) {
+        MONPUTSTR("Attacks up to 100 times.");
+    } else if (pm == &mons[PM_HYDRA]) {
+        MONPUTSTR("Attacks an extra time for each head beyond the first.");
+    }
     Sprintf(buf, "Is %sa valid polymorph form.",
             polyok(pm) ? "" : "not ");
     MONPUTSTR(buf);